First working version of chordpro.scm package definition
This packages the pango branch from the chordpro git repository.

This is my first guix package definition, and is therefore not  all
that great. There are a few things that aren't done the recommended
way, and will be revised once I learn them.

Contains quite a lot of comments for explanation. For myself.