Remove (propagated-inputs)
(propagated-inputs) seem to be unnecessary in this case. And also,
they seem to throw away the main goals of functional package mana-
gement in general: i.e. if A and B both depend on C (a propagated-
input: basically, propagated-inputs are actually installed in one'
s guix profile while inputs are not), A and B can't be upgraded i-
ndependently of each other.