* chordpro.scm: Bug Fix: Add missing dependencies
A few dependencies in the dependency chain were not available to
the program because of the fact that they were declared as prop-
agated-inputs. This was causing PDF rendering to fail, while txt
and html outputs were working pretty decently. This commit fixes
the issue by explicitly declaring 'the missing links' as inputs.