* chordpro.scm: Further simplify the 'wrap phase.
Thanks to suggestion by str1ngs from #guix on freenode for the idea

Instead of setting the PERL5LIB manually, just take the value of
PERL5LIB from the build environment. This is less error prone. And
a lot more 'functional'. The path of the master foo.

Also, change the wrapping type from 'prefix' to '='; because it is
enough that way.