* imglapse.scm: Simplify/polish the imglapse package definition
Since coreutils and sed are available in the build-system, just
use (getenv "PATH") in the 'wrap phase so as to have the program
access the required commands during runtime. While this does make
quite a few unneeded directories available in $PATH, this is still
cleaner and works just as well.